+++ lib/common/docker_functions.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package docker2aci
+package common
 
 import (
 	"strings"
