+++ lib/types/docker_types.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package docker2aci
+package types
 
 import "time"
