+++ lib/types/types.go
@@ -12,13 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package docker2aci
-
-type RepoData struct {
-	Tokens    []string
-	Endpoints []string
-	Cookie    []string
-}
+package types
 
 type ParsedDockerURL struct {
 	IndexURL  string
